Upgrade version, fix Windows build and add aarch64 Windows cross-compile (#2)
The Windows build was failing on `main` because of argument and path
mismatches: Visual Studio is a _multiconfig generator_ in `cmake`
and requires the build type to be set at *build* rather than *config* time,
otherwise it emits a `Debug` build by default.

Since we want to provide a native development experience on Windows
Aarch64 laptops (for example using the latest Qualcomm X Elite chips),
cross-compile an aarch64 library for Windows as well (note that
`macos-13-xlarge` is an Mx runner with `aarch64` chip already).  This
required a few workarounds to make sure `{clang,llvm}-tablegen.exe`
is *natively* compiled for the host which needs to run it, rather than
for the cross-compiled `aarch64` target.

Lastly, bump the hash to a commit from yesterday, and clean up some
of the scripts to not constantly be stuck with `cd` commands.